Support hbase op in video copy detection
